requirements.txt
addict==2.4.0
affine @ file:///Users/ktietz/demo/mc3/conda-bld/affine_1629890479180/work
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
aniso8601==9.0.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs @ file:///croot/attrs_1695717823297/work
augmentation==0.7
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bqplot==0.12.42
branca==0.7.0
cachelib==0.9.0
cachetools==5.3.2
category-encoders==2.6.3
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi==1.16.0
charset-normalizer==3.3.2
click @ file:///croot/click_1698129812380/work
click-plugins @ file:///Users/ktietz/demo/mc3/conda-bld/click-plugins_1630665883852/work
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.2.0
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
duckdb==0.9.2
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
fiona==1.9.5
Flask==2.3.3
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-restx==1.3.0
folium==0.15.1
fonttools==4.44.0
fqdn==1.5.1
fsspec==2023.12.2
future==0.18.3
GDAL==3.0.2
gdal2tiles==0.1.9
gdown==5.0.1
geojson==3.1.0
geopandas==0.13.2
groundingdino-py==0.4.0
h11==0.14.0
huggingface-hub==0.20.3
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.23.0
ipyleaflet==0.18.2
ipython==8.12.3
ipytree==0.2.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
large-image==1.27.1
large-image-source-rasterio==1.27.1
lazy_loader==0.3
leafmap==0.30.1
lightning-utilities==0.10.0
localtileserver==0.7.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmcv-full==1.7.1
-e git+https://github.com/open-mmlab/mmsegmentation.git@cbf9af1e3f2eb50b58c8d783008f749ef62f2435#egg=mmsegmentation
model-index==0.1.11
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.29
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.2
palettable==3.3.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
prettytable==3.9.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydensecrf @ file:///home/conda/feedstock_root/build_artifacts/pydensecrf_1695975708375/work
Pygments==2.16.1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyproj==3.5.0
pyshp==2.3.1
PySocks==1.7.1
pystac==1.8.4
pystac-client==0.7.5
python-box==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
qudida==0.0.4
rasterio==1.3.9
referencing==0.30.2
regex==2023.12.25
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.12.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scooby==0.9.2
seaborn==0.13.2
segment-anything-hq==0.3
segment-anything-py==1.0
segment-geospatial==0.10.2
Send2Trash==1.8.2
server-thread==0.2.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snuggs @ file:///Users/ktietz/demo/mc3/conda-bld/snuggs_1629803886966/work
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
supervision==0.6.0
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
torch==1.9.1+cu111
torchaudio==0.9.1
torcheval==0.0.7
torchmetrics==1.2.1
torchvision==0.10.1+cu111
tornado==6.3.3
tqdm==4.65.2
traitlets==5.13.0
traittypes==0.2.1
transformers==4.37.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.27.0.post1
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.9
xyzservices==2023.10.1
yapf==0.40.2
zipp==3.17.0
